Linting and converting to jsx
